Promote UiSurface to mod
